+++ CHANGELOG.md
@@ -12,7 +12,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/KDAB/cxx-qt/compare/v0.4.0...HEAD)
+## [Unreleased](https://github.com/KDAB/cxx-qt/compare/v0.4.1...HEAD)
+
+## [0.4.1](https://github.com/KDAB/cxx-qt/compare/v0.4.0...v0.4.1) - 2022-11-18
 
 ### Added